Adiciona tilemap e colisão do personagem
